backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase>=2.25.0
python-dotenv==1.0.0
pydantic>=2.12.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx>=0.28.0
email-validator>=2.0.0

